--- src/js/chat/leave.ts
@@ -1,13 +1,0 @@
-import { socket } from '../socket/socket';
-
-const leaveBtn = document.querySelector('#leave-btn');
-
-function leaveRoom(): void {
-    socket.emit('leaveRoom');
-}
-
-// 게임 나가기
-leaveBtn?.addEventListener('click', () => {
-    leaveRoom();
-    window.location.href = `/src/pages/room.html`;
-});
