+++ src/js/room/enterRoom.ts
@@ -1,15 +1,6 @@
 const urlParams = new URLSearchParams(window.location.search);
 const nickname = urlParams.get('nickname') as string;
 
-/**
- * 채팅방 입장 요청 파라미터를 정의하는 인터페이스
- */
-export interface JoinRoomParams {
-    roomId: string;
-    user_id: string;
-    nickName?: string;
-}
-
 // ✅ DOMContentLoaded 이벤트로 HTML 로딩 완료 후 실행 보장
 document.addEventListener('DOMContentLoaded', () => {
     const enterBtn = document.getElementById('enter-btn') as Element;
